requirements.txt
python-dotenv==1.0.1
cx-Oracle==8.3.0
Flask==3.0.2
requests==2.31.0
APScheduler==3.10.4
